feat(review): enhance aggregation logic by introducing AggregationState and updating method signatures to include missing mark mappings
